Move QualifiedFunctionNameWalker to Analysis project.
