+++ Python/Product/Debugger/Debugger/PythonStackFrame.cs
@@ -20,9 +20,8 @@ using System.IO;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
-using Microsoft.PythonTools.Infrastructure;
+using Microsoft.PythonTools.Analysis;
 using Microsoft.PythonTools.Parsing;
-using Microsoft.PythonTools.Parsing.Ast;
 
 namespace Microsoft.PythonTools.Debugger {
     class PythonStackFrame {
