+++ Python/Product/TestAdapter/PythonStackTraceParser.cs
@@ -18,10 +18,8 @@ using System;
 using System.Collections.Generic;
 using System.ComponentModel.Composition;
 using System.IO;
-using System.Text;
 using System.Text.RegularExpressions;
-using Microsoft.PythonTools.Debugger;
-using Microsoft.PythonTools.Infrastructure;
+using Microsoft.PythonTools.Analysis;
 using Microsoft.PythonTools.Parsing;
 using Microsoft.PythonTools.Parsing.Ast;
 using Microsoft.VisualStudio.TestWindow.Extensibility;
